move M_Inventory
https://github.com/metasfresh/metasfresh/issues/2647
