+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/inventory/interceptor/M_Inventory.java
@@ -1,4 +1,4 @@
-package de.metas.handlingunits.model.validator;
+package de.metas.handlingunits.inventory.interceptor;
 
 import static org.adempiere.model.InterfaceWrapperHelper.loadOutOfTrx;
